requirements.txt
altgraph==0.10.2
aniso8601==8.0.0
ansible==2.8.5
asn1crypto==0.24.0
backports.functools-lru-cache==1.6.1
bcrypt==3.1.4
bdist-mpkg==0.5.0
beautifulsoup4==4.8.2
bonjour-py==0.3
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
cvprac==0.9.0
docutils==0.14
enum34==1.1.6
Flask==1.1.1
Flask-RESTful==0.3.7
gitdb2==2.0.5
GitPython==2.1.11
idna==2.7
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.11.1
jsonrpclib==0.1.7
lxml==4.3.0
macholib==1.5.1
MarkupSafe==1.0
matplotlib==1.3.1
modulegraph==0.10.4
netaddr==0.7.19
netmiko==2.3.0
numpy==1.8.0rc1
paramiko==2.4.2
py2app==0.7.3
pyaml==17.12.1
pyasn1==0.4.4
pycparser==2.19
pycurl==7.43.0.3
pyeapi==0.8.2
pymongo==3.10.1
PyNaCl==1.3.0
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
pyserial==3.4
python-dateutil==1.5
pytz==2013.7
PyYAML==3.13
req==1.0.0
requests==2.20.0
scipy==0.13.0b1
scp==0.12.0
selenium==3.141.0
six==1.4.1
smmap2==2.0.5
soupsieve==1.9.5
textfsm==0.4.1
urllib3==1.24
virtualenv==16.1.0
Werkzeug==0.16.1
xattr==0.6.4
